All updated raghu

+++ backend/src/main/java/com/medi360/service/UserService.java
@@ -1,5 +1,6 @@
 package com.medi360.service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,25 +8,69 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import com.medi360.db.AuditlogRepository;
 import com.medi360.db.UserRepository;
-import com.medi360.entities.Patient;
+import com.medi360.entities.AuditLog;
 import com.medi360.entities.User;
+import com.medi360.exception.PatientNotFoundException;
+import com.medi360.exception.UserNotFoundException;
 
 @Service
 public class UserService {
 	
 	@Autowired
 	private UserRepository userrepo;
+	@Autowired
+	private AuditlogRepository auditrepo;
 	
+	public void addUserAuditLog(AuditLog au) {
+		this.auditrepo.save(au);
+	}
 	public User addUser(User user) {
+		
+		AuditLog au=new AuditLog();
+		au.setAction("User with Name "+user.getUserName()+" and Role "+user.getUserRole()+" successfully created");
+		au.setTimestamp(LocalDateTime.now());
+		au.setUser(user);
+		
+		addUserAuditLog(au);
 		return this.userrepo.save(user);
 	}
 	
-	public User updateUser(User user) {
+	public User updateUser(User user) throws UserNotFoundException{
+		if (!userrepo.existsById(user.getUserId())) {
+			throw new UserNotFoundException("User not found with id " + user.getUserId());
+		}
+		
+		AuditLog au=new AuditLog();
+		au.setAction("User with Name "+user.getUserName()+" and Role "+user.getUserRole()+" successfully updated");
+		au.setTimestamp(LocalDateTime.now());
+		au.setUser(user);
+		
+		addUserAuditLog(au);
+		
+		
+		
 		return this.userrepo.save(user);
+		
 	}
 	
-	public String deleteUser(int id) {
+	public String deleteUser(int id) throws UserNotFoundException{
+		if (!userrepo.existsById(id)) {
+			throw new UserNotFoundException("User not found with id " + id);
+		}
+		
+		User u=this.userrepo.findById(id).get();
+		
+		AuditLog au=new AuditLog();
+		au.setAction("User with Name "+u.getUserName()+" and Role "+u.getUserRole()+" successfully deleted");
+		au.setTimestamp(LocalDateTime.now());
+		au.setUser(u);
+		
+		addUserAuditLog(au);
+		
+		
+		
 		this.userrepo.deleteById(id);
 		return "Successfully Deleted";
 	}
@@ -37,4 +82,13 @@ public class UserService {
 	public Page<User> getAllUsersWithPagination(Pageable pageable) {
 		return this.userrepo.findAll(pageable);
 	}
+	
+	public User findById(int id) throws UserNotFoundException{
+		if (!userrepo.existsById(id)) {
+			throw new UserNotFoundException("User not found with id " + id);
+		}
+		return this.userrepo.findById(id).orElseThrow();
+	}
+	
+	
 }
